+++ changelog.md
@@ -1,6 +1,7 @@
 #### 1.0.5 (2021-02-15)
 
-- optimize fields (#89ac2ee)
+- Options for post types with archives (#f022303)
+- optimize fields (#b272529)
 - fix document title for posts (#6d7d4c2)
 
 #### 1.0.4 (2021-02-10)